+++ docs/src/plugin-zustand/changelog.md
@@ -11,3 +11,9 @@ outline: 2
 ## Next
 
 _Nothing yet._
+
+## 0.1.1
+
+### Features
+
+- Add [`try_state_or`](https://docs.rs/tauri-plugin-zustand/0.1.1/tauri_plugin_zustand/struct.Store.html#method.try_state_or), [`try_state_or_default`](https://docs.rs/tauri-plugin-zustand/0.1.1/tauri_plugin_zustand/struct.Store.html#method.try_state_or_default), and [`try_state_or_else`](https://docs.rs/tauri-plugin-zustand/0.1.1/tauri_plugin_zustand/struct.Store.html#method.try_state_or_else) methods.